UTL-351 Fix up Webpack 5 LESS compilation.
Bigger change than otherwise required because the library itself has been moved into subdirectory `library`, which is so we can verify what dependencies are needed in the test project - previously it would inherit modules from the parent directory which made it hard.

The main fix has been to turn off relative URL rewriting in less-loader, so any URLs to things like webfonts and images are emitted as-is. We currently rely on this behaviour to point at asset paths that we copy into place.